Fix: Properly handle ndbc files that don't have metadata
MHKiT-Python outputs a tuple from py.mhkit.wave.io.ndbc.read_file that
contains `(data, meta)`. For spectra, `meta` is set to None. Issue 94
details a warning that originates from this behavior. To fix this we
check to see if `meta` is `py.NoneType`. If yes we setup an empty
struct, which fixes the following warning:

"Warning: Calling STRUCT on an object prevents the object from hiding its
implementation details and should thus be avoided. Use DISP or DISPLAY to see
the visible public details of an object."